@@ -1,25 +1,24 @@
-﻿namespace Leetcode
+﻿namespace Leetcode.Logic;
+
+public class NumberOfBits
 {
-    public class NumberOfBits
+    public int HammingWeight(uint number)
     {
-        public int HammingWeight(uint number)
+        var sum = 0;
+
+        while (number > 0)
         {
-            var sum = 0;
+            if (number % 2 != 0)
+                sum++;
 
-            while (number > 0)
-            {
-                if (number % 2 != 0)
-                    sum++;
+            number /= 2;
+        };
 
-                number /= 2;
-            };
-            
-            return sum;
-        }
+        return sum;
+    }
 
-        public int HammingWeightString(uint number)
-        {
-            return Convert.ToString(number, 2).Count(bit => bit == '1');
-        }
+    public int HammingWeightString(uint number)
+    {
+        return Convert.ToString(number, 2).Count(bit => bit == '1');
     }
 }
