@@ -1,18 +1,18 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Leetcode.Logic;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Leetcode.Tests
+namespace LeetcodeTests;
+
+[TestClass()]
+public class ArithmeticProgressionFromSequenceTests
 {
-    [TestClass()]
-    public class ArithmeticProgressionFromSequenceTests
+    [TestMethod()]
+    public void CanMakeArithmeticProgressionTest()
     {
-        [TestMethod()]
-        public void CanMakeArithmeticProgressionTest()
-        {
-            var service = new ArithmeticProgressionFromSequence();
+        var service = new ArithmeticProgressionFromSequence();
 
-            Assert.AreEqual(true, service.CanMakeArithmeticProgression(new[] { 3, 5, 1 }));
+        Assert.AreEqual(true, service.CanMakeArithmeticProgression([3, 5, 1]));
 
-            Assert.AreEqual(false, service.CanMakeArithmeticProgression(new[] { 1, 2, 4 }));
-        }
+        Assert.AreEqual(false, service.CanMakeArithmeticProgression([1, 2, 4]));
     }
 }
