@@ -1,34 +1,34 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Leetcode;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Leetcode.Tests
+namespace LeetcodeTests;
+
+[TestClass()]
+public class OneStringSwapCanMakeStringsEqualTests
 {
-    [TestClass()]
-    public class OneStringSwapCanMakeStringsEqualTests
+    [TestMethod()]
+    public void AreAlmostEqualTest()
     {
-        [TestMethod()]
-        public void AreAlmostEqualTest()
-        {
-            var service = new OneStringSwapCanMakeStringsEqual();
+        var service = new OneStringSwapCanMakeStringsEqual();
 
-            Assert.AreEqual(true, service.AreAlmostEqual("bank", "kanb"));
+        Assert.AreEqual(true, service.AreAlmostEqual("bank", "kanb"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("attack", "defend"));
+        Assert.AreEqual(false, service.AreAlmostEqual("attack", "defend"));
 
-            Assert.AreEqual(true, service.AreAlmostEqual("kelb", "kelb"));
+        Assert.AreEqual(true, service.AreAlmostEqual("kelb", "kelb"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("abcd", "dcba"));
+        Assert.AreEqual(false, service.AreAlmostEqual("abcd", "dcba"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("caa", "aaz"));
+        Assert.AreEqual(false, service.AreAlmostEqual("caa", "aaz"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("npv", "zpn"));
+        Assert.AreEqual(false, service.AreAlmostEqual("npv", "zpn"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("bankb", "kannb"));
+        Assert.AreEqual(false, service.AreAlmostEqual("bankb", "kannb"));
 
-            Assert.AreEqual(true, service.AreAlmostEqual("aaXaaaXaaaaBaaa", "aaXaaaBaaaaXaaa"));
+        Assert.AreEqual(true, service.AreAlmostEqual("aaXaaaXaaaaBaaa", "aaXaaaBaaaaXaaa"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("aaXaaaXaaaaZaaa", "aaXaaaBaaaaXaaa"));
+        Assert.AreEqual(false, service.AreAlmostEqual("aaXaaaXaaaaZaaa", "aaXaaaBaaaaXaaa"));
 
-            Assert.AreEqual(false, service.AreAlmostEqual("aaXaaaXaaaaBaaa", "aaXaaaBaaaaZaaa"));
-        }
+        Assert.AreEqual(false, service.AreAlmostEqual("aaXaaaXaaaaBaaa", "aaXaaaBaaaaZaaa"));
     }
 }
