@@ -1,25 +1,24 @@
 ﻿using Leetcode;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Leetcode.Tests
+namespace LeetcodeTests;
+
+[TestClass()]
+public class LargestPerimeterTriangleTests
 {
-    [TestClass()]
-    public class LargestPerimeterTriangleTests
+    [TestMethod()]
+    public void LargestPerimeterTest()
     {
-        [TestMethod()]
-        public void LargestPerimeterTest()
-        {
-            var service = new LargestPerimeterTriangle();
+        var service = new LargestPerimeterTriangle();
 
-            Assert.AreEqual(5, service.LargestPerimeter(new[] { 2, 1, 2 }));
+        Assert.AreEqual(5, service.LargestPerimeter([2, 1, 2]));
 
-            Assert.AreEqual(0, service.LargestPerimeter(new[] { 1, 2, 1, 10 }));
+        Assert.AreEqual(0, service.LargestPerimeter([1, 2, 1, 10]));
 
-            Assert.AreEqual(10, service.LargestPerimeter(new[] { 3, 2, 3, 4 }));
+        Assert.AreEqual(10, service.LargestPerimeter([3, 2, 3, 4]));
 
-            Assert.AreEqual(361, service.LargestPerimeter(new[] { 4, 14, 73, 97, 80, 58, 12, 2, 2, 77, 98, 89, 26, 40, 49, 62, 65, 18, 9, 98, 10, 57, 27, 92, 28, 91, 93, 45, 42, 435, 41, 98, 30, 43, 19, 21, 88, 97, 4, 94, 22, 27, 28, 76, 20, 68, 15, 88, 77, 74, 74, 39, 64, 15, 6, 91, 15, 74, 72, 70, 51, 15, 67, 68, 89, 65, 81, 82, 44, 75, 17, 81, 83, 68, 62, 81, 21, 2, 70, 97, 74, 11, 3, 22, 7, 92, 7, 1, 60, 94, 91, 73, 72, 85, 67, 66, 35, 54215, 9, 53, 25, 83, 17, 9, 89, 18, 14, 32, 55, 37, 46, 97, 16, 31, 1148, 19, 57, 12, 93, 69, 10, 28, 79, 78, 90, 25, 9, 10, 29, 28, 47, 48, 40, 1, 50, 64, 70, 96, 70, 75, 74, 87, 24, 22, 2, 23, 27, 17, 63, 60, 3, 67, 16, 5, 72, 68, 32, 79, 45, 91, 8, 81, 85, 72, 100, 97, 32, 5, 12788, 17, 52, 15, 1, 60, 79, 44, 20, 81, 76, 55, 18, 265, 20, 2, 48, 30, 49, 62, 36, 49, 14, 30, 64, 18, 97, 38, 38, 51, 100, 33, 88, 27, 92, 18, 87, 22, 92, 64, 21, 50, 31, 23, 88, 78, 97, 21, 12, 95, 3, 40, 11, 46, 34, 86, 94, 55, 80, 57, 29, 92, 80, 37, 76, 47, 24, 19, 42, 99, 30, 23, 51, 83, 23, 30, 71, 35, 94, 90, 65, 83, 77, 47, 25, 9, 6, 100, 73, 68, 70, 82, 12, 53, 8, 37, 63, 50, 88, 79, 14, 54, 29, 27, 69, 100, 47, 53, 53, 85, 97, 55, 89, 1, 30, 50, 68, 100, 73, 30, 83, 40, 60, 7, 54, 100, 31, 21, 73, 98, 30, 85, 79, 63, 16, 66, 27, 26, 26, 35, 40, 87, 6, 34, 66, 6, 40, 40, 92, 11, 41, 35, 4, 17, 25, 59, 86, 88, 23, 84, 36, 21, 80, 75, 58, 8, 30, 17, 13, 44, 75, 46, 16, 75, 23, 47, 38, 68, 22, 37, 66, 92, 39, 43, 42, 95, 55, 43, 7, 49, 92, 20699, 58, 53, 25, 64, 75, 45, 94, 53, 16, 25, 77, 43, 39, 66, 5, 36, 67, 92, 8, 51, 7, 37, 29, 24, 62, 25, 65, 6, 6, 87, 88, 62, 76, 71, 4876, 79, 15, 81, 33, 93, 77, 35, 30, 94, 48, 89, 61, 98, 3, 81, 3, 43, 69, 47, 80, 80, 58, 18, 41, 94, 7, 4, 98, 14, 7, 61, 85, 23, 10, 27, 39, 90, 100, 61, 24, 89, 48, 92, 88, 1, 67, 91, 53, 50, 2, 24, 5, 60, 33, 87, 65, 74, 72, 12, 93, 25, 12, 76, 49, 85, 94, 34, 1857, 47, 100, 47, 30, 27, 61, 40, 34, 32, 40, 91, 97, 62, 45, 87712, 27, 6, 17, 76, 7, 21, 1, 15, 6, 61, 35, 45, 82, 28, 19, 75, 1, 92, 47, 50, 2, 32, 92, 20, 48, 92, 61, 45, 67, 69, 32, 16, 27, 39, 34, 73, 27, 85, 35, 72, 52, 73, 62, 23, 58, 41, 28, 58, 75, 70, 2, 33, 41, 25, 34, 39, 24, 83, 25, 50, 62, 92, 29, 99, 65, 24, 36, 75, 57, 35, 56, 29, 69, 90, 55, 71, 29, 90, 44, 11, 73, 78, 70, 87, 34, 65, 27, 33, 7, 91, 81, 85, 85, 61, 93, 29, 87, 14, 33496, 87, 54, 83, 64, 77, 39, 63, 39, 33, 46, 77, 23, 13, 96, 24, 95, 601290, 44, 74, 10, 72, 60, 72, 53, 84, 82, 3, 17, 56, 87, 56, 38, 50, 14, 91, 5, 24, 8, 7, 4, 35, 91, 34, 44, 68, 17, 82, 6, 98, 100, 100, 18, 5, 82, 3, 706, 47, 63, 42, 35, 12, 16, 13, 76, 75, 80, 82, 42, 141934, 94, 60, 27, 89, 4, 74, 43, 8, 5, 9, 42, 10, 17, 56, 93, 27, 32, 41, 37, 10, 48, 37, 81, 77, 18, 98, 15, 90, 21, 17, 93, 82, 92, 20, 80, 19, 3, 39, 80, 13, 65, 43, 76, 85, 92, 73, 14, 44, 93, 42, 61, 95, 50, 37, 71, 37, 46, 18, 79, 371611, 51, 40, 56, 62, 4, 3015, 66, 30, 60, 21, 53, 72, 1, 80, 33, 24, 77, 58, 78, 26, 71, 79, 73, 4, 37, 90, 70, 8, 94, 63, 29, 35, 21, 54, 39, 8, 86, 45, 54, 73, 61, 59, 19, 46, 96, 79, 31, 95, 52, 92, 48, 34, 43, 41, 33, 33, 52, 83, 83, 5, 54, 10, 64, 24, 40, 972911, 70, 161, 99, 45, 97, 229662, 77, 87, 65, 60, 70, 48, 17, 29, 45, 44, 83, 52, 56, 100, 52, 94, 73, 74, 46, 2, 37, 81, 26, 23, 93, 89, 41, 69, 2, 67, 66, 23, 79, 70, 5, 21, 7, 71, 91, 33, 69, 57, 25, 17, 55, 6, 9, 13, 60, 72, 57, 66, 26, 84, 56, 1, 96, 59, 75, 15, 61, 18, 76, 68, 23, 59, 40, 23, 41, 53, 20, 81, 6, 98, 91, 81, 62, 11, 39, 21, 7, 7902, 90, 93, 54, 39, 92, 31, 100, 4, 33, 73, 60, 49, 2, 9, 28, 49, 96, 30, 68, 49, 80, 33, 39, 91, 89, 99, 5, 83, 60, 34, 24, 5, 38, 72, 16, 29, 38, 25, 21, 66, 99, 19, 17, 95, 93, 89, 2, 73, 33, 34, 98, 17, 49, 48, 29, 86, 70, 82, 26, 78, 72, 6, 67, 98, 49, 17, 20, 46, 83, 37, 79, 25, 52, 47, 34, 19, 29, 76, 87, 51, 86, 36, 16, 4, 58, 88, 56, 69, 90, 45, 17, 42, 99, 34, 75, 79, 24, 80, 49, 45, 97, 38, 81, 5, 65, 56, 36, 63, 44, 55, 89, 15, 7, 34, 72, 47, 62, 32, 44, 50, 22, 34, 65, 68, 67, 69, 31, 13, 14, 19, 25, 83, 48, 61, 43, 67, 64, 56, 44, 91 }));
-            
-            Assert.AreEqual(2997877, service.LargestPerimeter(new[] { 606088, 610496, 425277, 487573, 380244, 74098, 4142, 570605, 100571, 118297, 424160, 33933, 828344, 144728, 674685, 152585, 971360, 449515, 628965, 927240, 173457, 49164, 383742, 993720, 21465, 455741, 269665, 210348, 977493, 70035, 418630, 760713, 203343, 935316, 514030, 474760, 176985, 822123, 706128, 800737, 434893, 574600, 436777, 565751, 260380, 739625, 997618, 318384, 916149, 984910, 115327, 806978, 162514, 457702, 43612, 735608, 388701, 370704, 608331, 388529, 910954, 680121, 200487, 894393, 465794, 200715, 845047, 57348, 529209, 856988, 19417, 155651, 387099, 300734, 45402, 492275, 611624, 469407, 85903, 912037, 196121, 97080, 419185, 476136, 964944, 644419, 217537, 472922, 530166, 107645, 969838, 81322, 731348, 566673, 33047, 365659, 395274, 113124, 651393, 262094, 2983, 434202, 8020, 912014, 537490, 220124, 84282, 606771, 146834, 72951, 133765, 343311, 473398, 662913, 595422, 606331, 805354, 968486, 575951, 880249, 239943, 250643, 440254, 568788, 882785, 658999, 55340, 516016, 764846, 690695, 417958, 978247, 132802, 506792, 237699, 177906, 898313, 721504, 52816, 452058, 309728, 342991, 610284, 830673, 74932, 493687, 763565, 746895, 302465, 309940, 310816, 125169, 293624, 703231, 779248, 602566, 874579, 886000, 196882, 969695, 285733, 964068, 263245, 187587, 840029, 938429, 949239, 892110, 311843, 582130, 212244, 801619, 763834, 32093, 745271, 507122, 601850, 222591, 979296, 253252, 978630, 601643, 346043, 298916, 914268, 605820, 114580, 354709, 649177, 767662, 252861, 101346, 334605, 360205, 800125, 417121, 61157, 638750, 234324, 584720, 294481, 146178, 497309, 7517, 940173, 391639, 685432, 242553, 827426, 443294, 328377, 649906, 443004, 735564, 396421, 417295, 718374, 27631, 94806, 920964, 530600, 669014, 87788, 215409, 594929, 799645, 628539, 409732, 497532, 342547, 557938, 543524, 455048, 471071, 847390, 755592, 685309, 958276, 357431, 129926, 916999, 822333, 697071, 857636, 719377, 169062, 230273, 26984, 333059, 793670, 803649, 268919, 897770, 256516, 88634, 146719, 909381, 895408, 92291, 592413, 526272, 32465, 512175, 417071, 419032, 99625, 445547, 420518, 920595, 421908, 536373, 965809, 921531, 42853, 160470, 959552, 18934, 610789, 82945, 103870, 833164, 187377, 209121, 510433, 429239, 531070, 419177, 599583, 619049, 510053, 10053, 518725, 895710, 800326, 594065, 625366, 887094, 199093, 513041, 383464, 37795, 357840, 875659, 253592, 705632, 213738, 457385, 519325, 623913, 933772, 360254, 723096, 439746, 895451, 733982, 775352, 446729, 602223, 497167, 708724, 253165, 955296, 65757, 465035, 353688, 452148, 439924, 776225, 432991, 278498, 609612, 631416, 515278, 279963, 15763, 93856, 104809, 148154, 971855, 463449, 145385, 751009, 388442, 617060, 187520, 609029, 942311, 976051, 570879, 902837, 504714, 376717, 937604, 955961, 693901, 746355, 896900, 356402, 398302, 78894, 612837, 292444, 731470, 664564, 907656, 275934, 4176, 283575, 456784, 743928, 594104, 395529, 63259, 232829, 665664, 993258, 800033, 287630, 245305, 646866, 828965, 53094, 321565, 393549, 246830, 777062, 243938, 533134, 44284, 92383, 969503, 596004, 744629, 642255, 652600, 331238, 119778, 884238, 402750, 802579, 656454, 550089, 635591, 526059, 396088, 836717, 228845, 521572, 916739, 593663, 275665, 830867, 284254, 920879, 199429, 498811, 17549, 96985, 438772, 403652, 597655, 568929, 998591, 298351, 329432, 257725, 735137, 318552, 313831, 299645, 32052, 467270, 844633, 771255, 493797, 59996, 374614, 432653, 833684, 436413, 380499, 70708, 914221, 894656, 648638, 326450, 393482, 594384, 28486, 647105, 755308, 921423, 580712, 108588, 123660, 697671, 432977, 37514, 849107, 106700, 593650, 441731, 12037, 535794, 720250, 691226, 947665, 730686, 401511, 978602, 17093, 845454, 347823, 834118, 736357, 151679, 564284, 448089, 802502, 879641, 715896, 673436, 690310, 233593, 561989, 199665, 219766, 808110, 177459, 968906, 877518, 229383, 938203, 888409, 547723, 353514, 915839, 496380, 681784, 610018, 338100, 409961, 88820, 367414, 34570, 978416, 402156, 341780, 841787, 367104, 69018, 851973, 226591, 647716, 46389, 698211, 462138, 145789, 953939, 436485, 591833, 482525, 225307, 86299, 578773, 227850, 649217, 939949, 256750, 902786, 950957, 154256, 463187, 754373, 828967, 999844, 152765, 560637, 462467, 546863, 450167, 526856, 859885, 261175, 354462, 962137, 338309, 439518, 460890, 672912, 881086, 786198, 734079, 353180, 22035, 372418, 429297, 38685, 468559, 223310, 54694, 794166, 663774, 210709, 648126, 93291, 544538, 389647, 728020, 673818, 261519, 975882, 45909, 77331, 128390, 895069, 972286, 588833, 129868, 719520, 153582, 12040, 710244, 749113, 980610, 66002, 86840, 70435, 562468, 396259, 760937, 930047, 723611, 251027, 37510, 226595, 69987, 177983, 776758, 382266, 101979, 18287, 884514, 865373, 653648, 256498, 986857, 307799, 457686, 546264, 826638, 458851, 875988, 348596, 788742, 239538, 827003, 51948, 632520, 544138, 338513, 420293, 858445, 619727, 312194, 826940, 377534, 536329, 964725, 851979, 535524, 776621, 376998, 278374, 326831, 158940, 723735, 766513, 128502, 942751, 318310, 911934, 112722, 156458, 268331, 61981, 392953, 247901, 437408, 894203, 544958, 52877, 196703, 215607, 922780, 219770, 332463, 10030, 621434, 110813, 106120, 102810, 403138, 829947, 423613, 439813, 728675, 919702, 10157, 563498, 816594, 368208, 11192, 159396, 68240, 768965, 82703, 502655, 781640, 163232, 24601, 3192, 326070, 709273, 589520, 390118, 256651, 870199, 297408, 957825, 770062, 723251, 649323, 410548, 313899, 121238, 348425, 96316, 363565, 91069, 35694, 930643, 352588, 486836, 801792, 275066, 405709, 279663, 455631, 663430, 784472, 791235, 174018, 281878, 817137, 536604, 558323, 163692, 463334, 388064, 838290, 281526, 941433, 648982, 714454, 899862, 943764, 341417, 930166, 375572, 702861, 364260, 289642, 251988, 334956, 465490, 560499, 971986, 260155, 82815, 620183, 658022, 989638, 255107, 997422, 441385, 397584, 95540, 337682, 518431, 141961, 995083, 472210, 303045, 401362, 825682, 369998, 910596, 623329, 403621, 156865, 765706, 10871, 22934, 871855, 243828, 303847, 383594, 125754, 292329, 679808, 48919, 6568, 684767, 958471, 589894, 849530, 405383, 284494, 382058, 799373, 877464, 993500, 533103, 97946, 910213, 807861, 7254, 836891, 419888, 999442, 787295, 788973, 188595, 50764, 853730, 854982, 377269, 735653, 942996, 21658, 2729, 124272, 941594, 361876, 669603, 842330, 278124, 258268, 979376, 462806, 608050, 322574, 190437, 877410, 824675, 795585, 405035, 718552, 992444, 511471, 40801, 864773, 825437, 825446, 113747, 248836, 288929, 410618, 264067, 303107, 578444, 402173, 644237, 809917, 809109, 138197, 533236, 618515, 184872, 8087, 321903, 833205, 402601, 963623, 968225, 715107, 441226, 119771, 683370, 360385, 286857, 859515, 220816, 641696, 350775, 189078, 705225, 610965, 406596, 164647, 674764, 352929, 591794, 751048, 613302, 97738, 646638, 986958, 939603, 535042, 959484, 947213, 911918, 255359, 397608, 730075, 875037, 881950, 197635, 499993, 614119, 967307, 247014, 415797, 308240, 145098, 465666, 885470, 102791, 782264, 12378, 199606, 92182, 30384, 283973, 166607, 397885, 513285, 801667, 320853, 229468, 865301, 49120, 133442, 773827, 2570, 70021, 808312, 767313, 235476, 611982, 411548, 393149, 347803, 750144, 659260, 458941, 671004, 222832, 195901, 218602, 17489, 812998, 634000, 830046, 229372, 598495, 775384, 667142, 666855, 147518, 578296, 323598, 215899, 680515, 403249, 469803, 77980, 563989, 258421, 257584, 108466, 251219, 248482, 416116, 974967, 343761, 262088, 562690, 355162, 743034, 488139, 626260, 22228, 95678, 755992, 99231, 155980, 481477, 89036, 425903, 242135, 70243, 464952, 804915, 604929, 445809, 975555, 832072, 295841, 572702, 28765, 982360, 174376, 415732, 142122, 496618, 733242, 464202, 265127, 755900, 134196, 923534, 402395, 80211, 375351, 578857, 972058, 342996, 22963, 947665, 892436, 810347, 811783, 680643, 332861, 37595, 659432, 216256 }));
-        }
+        Assert.AreEqual(361, service.LargestPerimeter([4, 14, 73, 97, 80, 58, 12, 2, 2, 77, 98, 89, 26, 40, 49, 62, 65, 18, 9, 98, 10, 57, 27, 92, 28, 91, 93, 45, 42, 435, 41, 98, 30, 43, 19, 21, 88, 97, 4, 94, 22, 27, 28, 76, 20, 68, 15, 88, 77, 74, 74, 39, 64, 15, 6, 91, 15, 74, 72, 70, 51, 15, 67, 68, 89, 65, 81, 82, 44, 75, 17, 81, 83, 68, 62, 81, 21, 2, 70, 97, 74, 11, 3, 22, 7, 92, 7, 1, 60, 94, 91, 73, 72, 85, 67, 66, 35, 54215, 9, 53, 25, 83, 17, 9, 89, 18, 14, 32, 55, 37, 46, 97, 16, 31, 1148, 19, 57, 12, 93, 69, 10, 28, 79, 78, 90, 25, 9, 10, 29, 28, 47, 48, 40, 1, 50, 64, 70, 96, 70, 75, 74, 87, 24, 22, 2, 23, 27, 17, 63, 60, 3, 67, 16, 5, 72, 68, 32, 79, 45, 91, 8, 81, 85, 72, 100, 97, 32, 5, 12788, 17, 52, 15, 1, 60, 79, 44, 20, 81, 76, 55, 18, 265, 20, 2, 48, 30, 49, 62, 36, 49, 14, 30, 64, 18, 97, 38, 38, 51, 100, 33, 88, 27, 92, 18, 87, 22, 92, 64, 21, 50, 31, 23, 88, 78, 97, 21, 12, 95, 3, 40, 11, 46, 34, 86, 94, 55, 80, 57, 29, 92, 80, 37, 76, 47, 24, 19, 42, 99, 30, 23, 51, 83, 23, 30, 71, 35, 94, 90, 65, 83, 77, 47, 25, 9, 6, 100, 73, 68, 70, 82, 12, 53, 8, 37, 63, 50, 88, 79, 14, 54, 29, 27, 69, 100, 47, 53, 53, 85, 97, 55, 89, 1, 30, 50, 68, 100, 73, 30, 83, 40, 60, 7, 54, 100, 31, 21, 73, 98, 30, 85, 79, 63, 16, 66, 27, 26, 26, 35, 40, 87, 6, 34, 66, 6, 40, 40, 92, 11, 41, 35, 4, 17, 25, 59, 86, 88, 23, 84, 36, 21, 80, 75, 58, 8, 30, 17, 13, 44, 75, 46, 16, 75, 23, 47, 38, 68, 22, 37, 66, 92, 39, 43, 42, 95, 55, 43, 7, 49, 92, 20699, 58, 53, 25, 64, 75, 45, 94, 53, 16, 25, 77, 43, 39, 66, 5, 36, 67, 92, 8, 51, 7, 37, 29, 24, 62, 25, 65, 6, 6, 87, 88, 62, 76, 71, 4876, 79, 15, 81, 33, 93, 77, 35, 30, 94, 48, 89, 61, 98, 3, 81, 3, 43, 69, 47, 80, 80, 58, 18, 41, 94, 7, 4, 98, 14, 7, 61, 85, 23, 10, 27, 39, 90, 100, 61, 24, 89, 48, 92, 88, 1, 67, 91, 53, 50, 2, 24, 5, 60, 33, 87, 65, 74, 72, 12, 93, 25, 12, 76, 49, 85, 94, 34, 1857, 47, 100, 47, 30, 27, 61, 40, 34, 32, 40, 91, 97, 62, 45, 87712, 27, 6, 17, 76, 7, 21, 1, 15, 6, 61, 35, 45, 82, 28, 19, 75, 1, 92, 47, 50, 2, 32, 92, 20, 48, 92, 61, 45, 67, 69, 32, 16, 27, 39, 34, 73, 27, 85, 35, 72, 52, 73, 62, 23, 58, 41, 28, 58, 75, 70, 2, 33, 41, 25, 34, 39, 24, 83, 25, 50, 62, 92, 29, 99, 65, 24, 36, 75, 57, 35, 56, 29, 69, 90, 55, 71, 29, 90, 44, 11, 73, 78, 70, 87, 34, 65, 27, 33, 7, 91, 81, 85, 85, 61, 93, 29, 87, 14, 33496, 87, 54, 83, 64, 77, 39, 63, 39, 33, 46, 77, 23, 13, 96, 24, 95, 601290, 44, 74, 10, 72, 60, 72, 53, 84, 82, 3, 17, 56, 87, 56, 38, 50, 14, 91, 5, 24, 8, 7, 4, 35, 91, 34, 44, 68, 17, 82, 6, 98, 100, 100, 18, 5, 82, 3, 706, 47, 63, 42, 35, 12, 16, 13, 76, 75, 80, 82, 42, 141934, 94, 60, 27, 89, 4, 74, 43, 8, 5, 9, 42, 10, 17, 56, 93, 27, 32, 41, 37, 10, 48, 37, 81, 77, 18, 98, 15, 90, 21, 17, 93, 82, 92, 20, 80, 19, 3, 39, 80, 13, 65, 43, 76, 85, 92, 73, 14, 44, 93, 42, 61, 95, 50, 37, 71, 37, 46, 18, 79, 371611, 51, 40, 56, 62, 4, 3015, 66, 30, 60, 21, 53, 72, 1, 80, 33, 24, 77, 58, 78, 26, 71, 79, 73, 4, 37, 90, 70, 8, 94, 63, 29, 35, 21, 54, 39, 8, 86, 45, 54, 73, 61, 59, 19, 46, 96, 79, 31, 95, 52, 92, 48, 34, 43, 41, 33, 33, 52, 83, 83, 5, 54, 10, 64, 24, 40, 972911, 70, 161, 99, 45, 97, 229662, 77, 87, 65, 60, 70, 48, 17, 29, 45, 44, 83, 52, 56, 100, 52, 94, 73, 74, 46, 2, 37, 81, 26, 23, 93, 89, 41, 69, 2, 67, 66, 23, 79, 70, 5, 21, 7, 71, 91, 33, 69, 57, 25, 17, 55, 6, 9, 13, 60, 72, 57, 66, 26, 84, 56, 1, 96, 59, 75, 15, 61, 18, 76, 68, 23, 59, 40, 23, 41, 53, 20, 81, 6, 98, 91, 81, 62, 11, 39, 21, 7, 7902, 90, 93, 54, 39, 92, 31, 100, 4, 33, 73, 60, 49, 2, 9, 28, 49, 96, 30, 68, 49, 80, 33, 39, 91, 89, 99, 5, 83, 60, 34, 24, 5, 38, 72, 16, 29, 38, 25, 21, 66, 99, 19, 17, 95, 93, 89, 2, 73, 33, 34, 98, 17, 49, 48, 29, 86, 70, 82, 26, 78, 72, 6, 67, 98, 49, 17, 20, 46, 83, 37, 79, 25, 52, 47, 34, 19, 29, 76, 87, 51, 86, 36, 16, 4, 58, 88, 56, 69, 90, 45, 17, 42, 99, 34, 75, 79, 24, 80, 49, 45, 97, 38, 81, 5, 65, 56, 36, 63, 44, 55, 89, 15, 7, 34, 72, 47, 62, 32, 44, 50, 22, 34, 65, 68, 67, 69, 31, 13, 14, 19, 25, 83, 48, 61, 43, 67, 64, 56, 44, 91]));
+
+        Assert.AreEqual(2997877, service.LargestPerimeter([606088, 610496, 425277, 487573, 380244, 74098, 4142, 570605, 100571, 118297, 424160, 33933, 828344, 144728, 674685, 152585, 971360, 449515, 628965, 927240, 173457, 49164, 383742, 993720, 21465, 455741, 269665, 210348, 977493, 70035, 418630, 760713, 203343, 935316, 514030, 474760, 176985, 822123, 706128, 800737, 434893, 574600, 436777, 565751, 260380, 739625, 997618, 318384, 916149, 984910, 115327, 806978, 162514, 457702, 43612, 735608, 388701, 370704, 608331, 388529, 910954, 680121, 200487, 894393, 465794, 200715, 845047, 57348, 529209, 856988, 19417, 155651, 387099, 300734, 45402, 492275, 611624, 469407, 85903, 912037, 196121, 97080, 419185, 476136, 964944, 644419, 217537, 472922, 530166, 107645, 969838, 81322, 731348, 566673, 33047, 365659, 395274, 113124, 651393, 262094, 2983, 434202, 8020, 912014, 537490, 220124, 84282, 606771, 146834, 72951, 133765, 343311, 473398, 662913, 595422, 606331, 805354, 968486, 575951, 880249, 239943, 250643, 440254, 568788, 882785, 658999, 55340, 516016, 764846, 690695, 417958, 978247, 132802, 506792, 237699, 177906, 898313, 721504, 52816, 452058, 309728, 342991, 610284, 830673, 74932, 493687, 763565, 746895, 302465, 309940, 310816, 125169, 293624, 703231, 779248, 602566, 874579, 886000, 196882, 969695, 285733, 964068, 263245, 187587, 840029, 938429, 949239, 892110, 311843, 582130, 212244, 801619, 763834, 32093, 745271, 507122, 601850, 222591, 979296, 253252, 978630, 601643, 346043, 298916, 914268, 605820, 114580, 354709, 649177, 767662, 252861, 101346, 334605, 360205, 800125, 417121, 61157, 638750, 234324, 584720, 294481, 146178, 497309, 7517, 940173, 391639, 685432, 242553, 827426, 443294, 328377, 649906, 443004, 735564, 396421, 417295, 718374, 27631, 94806, 920964, 530600, 669014, 87788, 215409, 594929, 799645, 628539, 409732, 497532, 342547, 557938, 543524, 455048, 471071, 847390, 755592, 685309, 958276, 357431, 129926, 916999, 822333, 697071, 857636, 719377, 169062, 230273, 26984, 333059, 793670, 803649, 268919, 897770, 256516, 88634, 146719, 909381, 895408, 92291, 592413, 526272, 32465, 512175, 417071, 419032, 99625, 445547, 420518, 920595, 421908, 536373, 965809, 921531, 42853, 160470, 959552, 18934, 610789, 82945, 103870, 833164, 187377, 209121, 510433, 429239, 531070, 419177, 599583, 619049, 510053, 10053, 518725, 895710, 800326, 594065, 625366, 887094, 199093, 513041, 383464, 37795, 357840, 875659, 253592, 705632, 213738, 457385, 519325, 623913, 933772, 360254, 723096, 439746, 895451, 733982, 775352, 446729, 602223, 497167, 708724, 253165, 955296, 65757, 465035, 353688, 452148, 439924, 776225, 432991, 278498, 609612, 631416, 515278, 279963, 15763, 93856, 104809, 148154, 971855, 463449, 145385, 751009, 388442, 617060, 187520, 609029, 942311, 976051, 570879, 902837, 504714, 376717, 937604, 955961, 693901, 746355, 896900, 356402, 398302, 78894, 612837, 292444, 731470, 664564, 907656, 275934, 4176, 283575, 456784, 743928, 594104, 395529, 63259, 232829, 665664, 993258, 800033, 287630, 245305, 646866, 828965, 53094, 321565, 393549, 246830, 777062, 243938, 533134, 44284, 92383, 969503, 596004, 744629, 642255, 652600, 331238, 119778, 884238, 402750, 802579, 656454, 550089, 635591, 526059, 396088, 836717, 228845, 521572, 916739, 593663, 275665, 830867, 284254, 920879, 199429, 498811, 17549, 96985, 438772, 403652, 597655, 568929, 998591, 298351, 329432, 257725, 735137, 318552, 313831, 299645, 32052, 467270, 844633, 771255, 493797, 59996, 374614, 432653, 833684, 436413, 380499, 70708, 914221, 894656, 648638, 326450, 393482, 594384, 28486, 647105, 755308, 921423, 580712, 108588, 123660, 697671, 432977, 37514, 849107, 106700, 593650, 441731, 12037, 535794, 720250, 691226, 947665, 730686, 401511, 978602, 17093, 845454, 347823, 834118, 736357, 151679, 564284, 448089, 802502, 879641, 715896, 673436, 690310, 233593, 561989, 199665, 219766, 808110, 177459, 968906, 877518, 229383, 938203, 888409, 547723, 353514, 915839, 496380, 681784, 610018, 338100, 409961, 88820, 367414, 34570, 978416, 402156, 341780, 841787, 367104, 69018, 851973, 226591, 647716, 46389, 698211, 462138, 145789, 953939, 436485, 591833, 482525, 225307, 86299, 578773, 227850, 649217, 939949, 256750, 902786, 950957, 154256, 463187, 754373, 828967, 999844, 152765, 560637, 462467, 546863, 450167, 526856, 859885, 261175, 354462, 962137, 338309, 439518, 460890, 672912, 881086, 786198, 734079, 353180, 22035, 372418, 429297, 38685, 468559, 223310, 54694, 794166, 663774, 210709, 648126, 93291, 544538, 389647, 728020, 673818, 261519, 975882, 45909, 77331, 128390, 895069, 972286, 588833, 129868, 719520, 153582, 12040, 710244, 749113, 980610, 66002, 86840, 70435, 562468, 396259, 760937, 930047, 723611, 251027, 37510, 226595, 69987, 177983, 776758, 382266, 101979, 18287, 884514, 865373, 653648, 256498, 986857, 307799, 457686, 546264, 826638, 458851, 875988, 348596, 788742, 239538, 827003, 51948, 632520, 544138, 338513, 420293, 858445, 619727, 312194, 826940, 377534, 536329, 964725, 851979, 535524, 776621, 376998, 278374, 326831, 158940, 723735, 766513, 128502, 942751, 318310, 911934, 112722, 156458, 268331, 61981, 392953, 247901, 437408, 894203, 544958, 52877, 196703, 215607, 922780, 219770, 332463, 10030, 621434, 110813, 106120, 102810, 403138, 829947, 423613, 439813, 728675, 919702, 10157, 563498, 816594, 368208, 11192, 159396, 68240, 768965, 82703, 502655, 781640, 163232, 24601, 3192, 326070, 709273, 589520, 390118, 256651, 870199, 297408, 957825, 770062, 723251, 649323, 410548, 313899, 121238, 348425, 96316, 363565, 91069, 35694, 930643, 352588, 486836, 801792, 275066, 405709, 279663, 455631, 663430, 784472, 791235, 174018, 281878, 817137, 536604, 558323, 163692, 463334, 388064, 838290, 281526, 941433, 648982, 714454, 899862, 943764, 341417, 930166, 375572, 702861, 364260, 289642, 251988, 334956, 465490, 560499, 971986, 260155, 82815, 620183, 658022, 989638, 255107, 997422, 441385, 397584, 95540, 337682, 518431, 141961, 995083, 472210, 303045, 401362, 825682, 369998, 910596, 623329, 403621, 156865, 765706, 10871, 22934, 871855, 243828, 303847, 383594, 125754, 292329, 679808, 48919, 6568, 684767, 958471, 589894, 849530, 405383, 284494, 382058, 799373, 877464, 993500, 533103, 97946, 910213, 807861, 7254, 836891, 419888, 999442, 787295, 788973, 188595, 50764, 853730, 854982, 377269, 735653, 942996, 21658, 2729, 124272, 941594, 361876, 669603, 842330, 278124, 258268, 979376, 462806, 608050, 322574, 190437, 877410, 824675, 795585, 405035, 718552, 992444, 511471, 40801, 864773, 825437, 825446, 113747, 248836, 288929, 410618, 264067, 303107, 578444, 402173, 644237, 809917, 809109, 138197, 533236, 618515, 184872, 8087, 321903, 833205, 402601, 963623, 968225, 715107, 441226, 119771, 683370, 360385, 286857, 859515, 220816, 641696, 350775, 189078, 705225, 610965, 406596, 164647, 674764, 352929, 591794, 751048, 613302, 97738, 646638, 986958, 939603, 535042, 959484, 947213, 911918, 255359, 397608, 730075, 875037, 881950, 197635, 499993, 614119, 967307, 247014, 415797, 308240, 145098, 465666, 885470, 102791, 782264, 12378, 199606, 92182, 30384, 283973, 166607, 397885, 513285, 801667, 320853, 229468, 865301, 49120, 133442, 773827, 2570, 70021, 808312, 767313, 235476, 611982, 411548, 393149, 347803, 750144, 659260, 458941, 671004, 222832, 195901, 218602, 17489, 812998, 634000, 830046, 229372, 598495, 775384, 667142, 666855, 147518, 578296, 323598, 215899, 680515, 403249, 469803, 77980, 563989, 258421, 257584, 108466, 251219, 248482, 416116, 974967, 343761, 262088, 562690, 355162, 743034, 488139, 626260, 22228, 95678, 755992, 99231, 155980, 481477, 89036, 425903, 242135, 70243, 464952, 804915, 604929, 445809, 975555, 832072, 295841, 572702, 28765, 982360, 174376, 415732, 142122, 496618, 733242, 464202, 265127, 755900, 134196, 923534, 402395, 80211, 375351, 578857, 972058, 342996, 22963, 947665, 892436, 810347, 811783, 680643, 332861, 37595, 659432, 216256]));
     }
 }
